Merge
Former-commit-id: bb347cb2da6c7a6569bd749c6acd9bb91372e5c0

+++ StRoot/KFParticle/KFPInputData.h
@@ -25,7 +25,7 @@ using std::ifstream;
 
 struct KFPTrackIndex
 {
-  short fIndex;
+  int fIndex;
   short fPdg;
   
   static bool Compare(const KFPTrackIndex& a, const KFPTrackIndex& b)
